style(components): move `ConnectButton` to outer scope
