--- app/components/connect/index.ts
@@ -1,1 +1,0 @@
-export * from './button'
